Moved vehicle respawn handling to server side. Documented vehicle damage information in AUTOMOBILE_TYPE structure. Fixed commands sending when not connected. Added CNetGame::Connect, CNetGame::IsConnected and CNetGame::Disconnect. Fixed CPlayerPed::GetCurrentWeapon returning the current weapon slot instead of the current weapon type.

+++ vcmp/game/playerped.h
@@ -99,7 +99,7 @@ public:
 	BOOL           GiveWeapon(int iWeaponID, int iAmmo);
 	void           ClearAllWeapons();
 	BOOL           SetArmedWeapon(int iWeaponType);
-	BYTE           GetCurrentWeapon();
+	int            GetCurrentWeapon();
 	BOOL           HasAmmoForCurrentWeapon();
 	WEAPON_SLOT *  GetCurrentWeaponSlot();
 	WEAPON_SLOT *  GetWeaponInSlot(BYTE byteSlot);
